% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aorisitc.R
\name{aoristic_server}
\alias{aoristic_server}
\title{Aoristic Analysis Server}
\usage{
aoristic_server(id, x, y)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{list} (returned by \code{\link[=time_interval_server]{time_interval_server()}})}
}
\value{
A reactive \code{\link[kairos:AoristicSum-class]{kairos::AoristicSum}} object.
}
\description{
Aoristic Analysis Server
}
\seealso{
\code{\link[=aoristic_ui]{aoristic_ui()}}

Other chronology modules: 
\code{\link{aoristic_ui}()},
\code{\link{seriate_server}()},
\code{\link{seriate_ui}()},
\code{\link{time_interval_server}()},
\code{\link{time_interval_ui}()}
}
\concept{chronology modules}
\keyword{internal}
