% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aorisitc.R
\name{aoristic_ui}
\alias{aoristic_ui}
\title{Aoristic Analysis UI}
\usage{
aoristic_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\value{
A nav item that may be passed to a nav container
(e.g. \code{\link[bslib:navset]{bslib::navset_tab()}}).
}
\description{
Aoristic Analysis UI
}
\seealso{
\code{\link[=aoristic_server]{aoristic_server()}}

Other chronology modules: 
\code{\link{aoristic_server}()},
\code{\link{seriate_server}()},
\code{\link{seriate_ui}()},
\code{\link{time_interval_server}()},
\code{\link{time_interval_ui}()}
}
\concept{chronology modules}
\keyword{internal}
