% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bertin.R
\name{bertin_ui}
\alias{bertin_ui}
\title{Bertin Diagrams UI}
\usage{
bertin_ui(id, title = NULL)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}

\item{title}{A \code{\link{character}} string giving the plot title.}
}
\value{
A nav item that may be passed to a nav container
(e.g. \code{\link[bslib:navset]{bslib::navset_tab()}}).
}
\description{
Bertin Diagrams UI
}
\seealso{
\code{\link[=bertin_server]{bertin_server()}}

Other count data modules: 
\code{\link{bertin_server}()},
\code{\link{count_server}()},
\code{\link{count_ui}()},
\code{\link{diversity_alpha_server}()},
\code{\link{diversity_alpha_ui}()},
\code{\link{diversity_beta_server}()},
\code{\link{diversity_beta_ui}()},
\code{\link{diversity_docs_ui}()},
\code{\link{diversity_server}()},
\code{\link{diversity_ui}()},
\code{\link{occurrence_server}()},
\code{\link{occurrence_ui}()}
}
\concept{count data modules}
\keyword{internal}
