% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca.R
\name{ca_server}
\alias{ca_server}
\title{Correspondence Analysis Server}
\usage{
ca_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{data.frame}.}
}
\value{
A reactive \code{\link[dimensio:PCA-class]{dimensio::PCA}} object.
}
\description{
Correspondence Analysis Server
}
\seealso{
\code{\link[=ca_ui]{ca_ui()}}

Other multivariate analysis modules: 
\code{\link{ca_ui}()},
\code{\link{multivariate_server}()},
\code{\link{multivariate_ui}()},
\code{\link{pca_server}()},
\code{\link{pca_ui}()}
}
\concept{multivariate analysis modules}
\keyword{internal}
