% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca.R
\name{ca_ui}
\alias{ca_ui}
\title{Correspondence Analysis UI}
\usage{
ca_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Correspondence Analysis UI
}
\seealso{
\code{\link[=ca_server]{ca_server()}}

Other multivariate analysis modules: 
\code{\link{ca_server}()},
\code{\link{multivariate_server}()},
\code{\link{multivariate_ui}()},
\code{\link{pca_server}()},
\code{\link{pca_ui}()}
}
\concept{multivariate analysis modules}
\keyword{internal}
