% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate_server}
\alias{calibrate_server}
\title{Radiocarbon Calibration Server}
\usage{
calibrate_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{data.frame} (typically returned by \code{\link[=import_server]{import_server()}}).}
}
\value{
A reactive \code{\link[ananke:CalibratedAges-class]{ananke::CalibratedAges}} object.
}
\description{
Radiocarbon Calibration Server
}
\seealso{
\code{\link[=calibrate_ui]{calibrate_ui()}}

Other chronology modules:
\code{\link[=aoristic_server]{aoristic_server()}},
\code{\link[=aoristic_ui]{aoristic_ui()}},
\code{\link[=calibrate_ui]{calibrate_ui()}},
\code{\link[=seriate_server]{seriate_server()}},
\code{\link[=seriate_ui]{seriate_ui()}},
\code{\link[=time_interval_server]{time_interval_server()}},
\code{\link[=time_interval_ui]{time_interval_ui()}}
}
\concept{chronology modules}
\keyword{internal}
