% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate_ui}
\alias{calibrate_ui}
\title{Radiocarbon Calibration UI}
\usage{
calibrate_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\value{
A nav item that may be passed to a nav container
(e.g. \code{\link[bslib:navset_tab]{bslib::navset_tab()}}).
}
\description{
Radiocarbon Calibration UI
}
\seealso{
\code{\link[=calibrate_server]{calibrate_server()}}

Other chronology modules:
\code{\link[=aoristic_server]{aoristic_server()}},
\code{\link[=aoristic_ui]{aoristic_ui()}},
\code{\link[=calibrate_server]{calibrate_server()}},
\code{\link[=seriate_server]{seriate_server()}},
\code{\link[=seriate_ui]{seriate_ui()}},
\code{\link[=time_interval_server]{time_interval_server()}},
\code{\link[=time_interval_ui]{time_interval_ui()}}
}
\concept{chronology modules}
\keyword{internal}
