% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{checkbox_ui}
\alias{checkbox_ui}
\title{Updatable Checkbox Group}
\usage{
checkbox_ui(id, label = "Choose", inline = TRUE)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\value{
A checkbox group control that can be added to a UI definition
(see \code{\link[shiny:checkboxGroupInput]{shiny::checkboxGroupInput()}}).
}
\description{
Updatable Checkbox Group
}
\seealso{
\code{\link[=update_checkbox_colnames]{update_checkbox_colnames()}}
}
\keyword{internal}
