% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi2.R
\name{chi2_server}
\alias{chi2_server}
\title{Chi-squared Test Server}
\usage{
chi2_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{data.frame} (typically returned by \code{\link[=import_server]{import_server()}}).}
}
\description{
Chi-squared Test Server
}
\seealso{
\code{\link[=chi2_ui]{chi2_ui()}}

Other statistic modules: 
\code{\link{chi2_ui}()}
}
\concept{statistic modules}
\keyword{internal}
