% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi2.R
\name{chi2_ui}
\alias{chi2_ui}
\title{Chi-squared Test UI}
\usage{
chi2_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Chi-squared Test UI
}
\seealso{
\code{\link[=chi2_server]{chi2_server()}}

Other statistic modules: 
\code{\link{chi2_server}()}
}
\concept{statistic modules}
\keyword{internal}
