% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_hclust.R
\name{coda_hclust_server}
\alias{coda_hclust_server}
\title{Compositional Data Hierarchical Clustering Server}
\usage{
coda_hclust_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{\link[nexus:CompositionMatrix-class]{nexus::CompositionMatrix}} object.}
}
\value{
No return value, called for side effects.
}
\description{
Compositional Data Hierarchical Clustering Server
}
\seealso{
\code{\link[=coda_hclust_ui]{coda_hclust_ui()}}

Other coda modules: 
\code{\link{coda_barplot_server}()},
\code{\link{coda_barplot_ui}()},
\code{\link{coda_hclust_ui}()},
\code{\link{coda_outliers_server}()},
\code{\link{coda_outliers_ui}()},
\code{\link{coda_server}()},
\code{\link{coda_summary_server}()},
\code{\link{coda_summary_ui}()},
\code{\link{coda_ui}()},
\code{\link{logratio_server}()},
\code{\link{logratio_ui}()}
}
\concept{coda modules}
\keyword{internal}
