% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_hclust.R
\name{coda_hclust_ui}
\alias{coda_hclust_ui}
\title{Compositional Data Hierarchical Clustering UI}
\usage{
coda_hclust_ui(id)
}
\description{
Compositional Data Hierarchical Clustering UI
}
\seealso{
\code{\link[=coda_hclust_server]{coda_hclust_server()}}

Other coda modules: 
\code{\link{coda_hclust_server}()},
\code{\link{coda_outliers_server}()},
\code{\link{coda_outliers_ui}()},
\code{\link{coda_plot_server}()},
\code{\link{coda_plot_ui}()},
\code{\link{coda_server}()},
\code{\link{coda_summary_server}()},
\code{\link{coda_summary_ui}()},
\code{\link{coda_ui}()},
\code{\link{logratio_server}()},
\code{\link{logratio_ui}()}
}
\concept{coda modules}
\keyword{internal}
