% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_coda_outliers.R
\name{coda_outliers_server}
\alias{coda_outliers_server}
\title{Compositional Data Outliers Server}
\usage{
coda_outliers_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{\link[nexus:CompositionMatrix-class]{nexus::CompositionMatrix}} object.}
}
\value{
A reactive \code{\link[nexus:OutlierIndex-class]{nexus::OutlierIndex}} object.
}
\description{
Compositional Data Outliers Server
}
\seealso{
\code{\link[=coda_outliers_ui]{coda_outliers_ui()}}

Other coda modules: 
\code{\link{coda_outliers_ui}()},
\code{\link{coda_plot_server}()},
\code{\link{coda_plot_ui}()},
\code{\link{coda_server}()},
\code{\link{coda_summary_server}()},
\code{\link{coda_summary_ui}()},
\code{\link{coda_ui}()},
\code{\link{logratio_server}()},
\code{\link{logratio_ui}()}
}
\concept{coda modules}
\keyword{internal}
