% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_outliers.R
\name{coda_outliers_ui}
\alias{coda_outliers_ui}
\title{Compositional Data Outliers UI}
\usage{
coda_outliers_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\value{
A nav item that may be passed to a nav container
(e.g. \code{\link[bslib:navset_tab]{bslib::navset_tab()}}).
}
\description{
Compositional Data Outliers UI
}
\seealso{
\code{\link[=coda_outliers_server]{coda_outliers_server()}}

Other coda modules:
\code{\link[=coda_barplot_server]{coda_barplot_server()}},
\code{\link[=coda_barplot_ui]{coda_barplot_ui()}},
\code{\link[=coda_hclust_server]{coda_hclust_server()}},
\code{\link[=coda_hclust_ui]{coda_hclust_ui()}},
\code{\link[=coda_hist_server]{coda_hist_server()}},
\code{\link[=coda_hist_ui]{coda_hist_ui()}},
\code{\link{coda_nav}},
\code{\link[=coda_outliers_server]{coda_outliers_server()}},
\code{\link[=coda_server]{coda_server()}},
\code{\link[=coda_summary_server]{coda_summary_server()}},
\code{\link[=coda_summary_ui]{coda_summary_ui()}},
\code{\link[=coda_ui]{coda_ui()}},
\code{\link[=logratio_server]{logratio_server()}},
\code{\link[=logratio_ui]{logratio_ui()}}
}
\concept{coda modules}
\keyword{internal}
