% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_coda_plot.R
\name{coda_plot_ui}
\alias{coda_plot_ui}
\title{Bar Plot UI}
\usage{
coda_plot_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Bar Plot UI
}
\seealso{
\code{\link[=coda_plot_server]{coda_plot_server()}}

Other coda modules: 
\code{\link{coda_outliers_server}()},
\code{\link{coda_outliers_ui}()},
\code{\link{coda_plot_server}()},
\code{\link{coda_server}()},
\code{\link{coda_summary_server}()},
\code{\link{coda_summary_ui}()},
\code{\link{coda_ui}()},
\code{\link{logratio_server}()},
\code{\link{logratio_ui}()}
}
\concept{coda modules}
\keyword{internal}
