% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.R
\name{coda_server}
\alias{coda_server}
\title{Compositional Data Server}
\usage{
coda_server(id, demo = NULL, verbose = get_option("verbose", FALSE))
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{demo}{A \code{\link{character}} string specifying the name of a dataset (see
\code{\link[=import_server]{import_server()}}).}

\item{verbose}{A \code{\link{logical}} scalar: should \R report extra information
on progress?}
}
\value{
A reactive \code{\link[nexus:CompositionMatrix-class]{nexus::CompositionMatrix}} object.
}
\description{
Compositional Data Server
}
\seealso{
\code{\link[=coda_ui]{coda_ui()}}

Other coda modules: 
\code{\link{coda_barplot_server}()},
\code{\link{coda_barplot_ui}()},
\code{\link{coda_hclust_server}()},
\code{\link{coda_hclust_ui}()},
\code{\link{coda_outliers_server}()},
\code{\link{coda_outliers_ui}()},
\code{\link{coda_summary_server}()},
\code{\link{coda_summary_ui}()},
\code{\link{coda_ui}()},
\code{\link{logratio_server}()},
\code{\link{logratio_ui}()}
}
\concept{coda modules}
\keyword{internal}
