% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_coda.R
\name{coda_ui}
\alias{coda_ui}
\title{Compositional Data UI}
\usage{
coda_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Compositional Data UI
}
\seealso{
\code{\link[=coda_server]{coda_server()}}

Other coda modules: 
\code{\link{coda_plot_server}()},
\code{\link{coda_plot_ui}()},
\code{\link{coda_server}()},
\code{\link{coda_summary_server}()},
\code{\link{coda_summary_ui}()},
\code{\link{logratio_server}()},
\code{\link{logratio_ui}()}
}
\concept{coda modules}
\keyword{internal}
