% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis-internal.R
\name{compute_server}
\alias{compute_server}
\title{Compute on Action Server}
\usage{
compute_server(id, x, f)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive object.}

\item{f}{A \code{\link{function}} that takes \code{x} as argument.}
}
\value{
A reactive object (the result of \code{f(x)}).
}
\description{
Compute on Action Server
}
\seealso{
\code{\link[=compute_ui]{compute_ui()}}
}
\keyword{internal}
