% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis-internal.R
\name{compute_ui}
\alias{compute_ui}
\title{Compute on Action UI}
\usage{
compute_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Compute on Action UI
}
\seealso{
\code{\link[=compute_server]{compute_server()}}
}
\keyword{internal}
