% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_count.R
\name{count_server}
\alias{count_server}
\title{Count Data Server}
\usage{
count_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{data.frame} (typically returned by \code{\link[=data_server]{data_server()}}).}
}
\value{
A reactive \code{\link{data.frame}}.
}
\description{
Count Data Server
}
\concept{count data modules}
\keyword{internal}
