% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_ui}
\alias{count_ui}
\title{Count Data UI}
\usage{
count_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\value{
A nav item that may be passed to a nav container
(e.g. \code{\link[bslib:navset]{bslib::navset_tab()}}).
}
\description{
Count Data UI
}
\seealso{
\code{\link[=count_server]{count_server()}}

Other count data modules: 
\code{\link{bertin_server}()},
\code{\link{bertin_ui}()},
\code{\link{count_server}()},
\code{\link{diversity_alpha_server}()},
\code{\link{diversity_alpha_ui}()},
\code{\link{diversity_beta_server}()},
\code{\link{diversity_beta_ui}()},
\code{\link{diversity_docs_ui}()},
\code{\link{diversity_server}()},
\code{\link{diversity_ui}()},
\code{\link{occurrence_server}()},
\code{\link{occurrence_ui}()}
}
\concept{count data modules}
\keyword{internal}
