% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data.R
\name{data_server}
\alias{data_server}
\title{Data Server}
\usage{
data_server(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
A reactive \code{data.frame}.
}
\description{
Data Server
}
\seealso{
\code{\link[=data_ui]{data_ui()}}

Other generic modules: 
\code{\link{data_ui}()},
\code{\link{import_server}()},
\code{\link{import_ui}()},
\code{\link{missing_server}()},
\code{\link{missing_ui}()},
\code{\link{prepare_server}()},
\code{\link{prepare_ui}()}
}
\concept{generic modules}
\keyword{internal}
