% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data.R
\name{data_ui}
\alias{data_ui}
\title{Data UI}
\usage{
data_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Data UI
}
\seealso{
\code{\link[=data_server]{data_server()}}

Other generic modules: 
\code{\link{data_server}()},
\code{\link{import_server}()},
\code{\link{import_ui}()},
\code{\link{missing_server}()},
\code{\link{missing_ui}()},
\code{\link{prepare_server}()},
\code{\link{prepare_ui}()}
}
\concept{generic modules}
\keyword{internal}
