% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy}
\alias{deploy}
\title{Deploy an App}
\usage{
deploy(app = c("seriate", "ca"))
}
\arguments{
\item{app}{A \code{\link{character}} string specifying the Shiny application
to run (see details). Any unambiguous substring can be given.}
}
\description{
Deploy a shiny application with \pkg{rsconnect}.
}
\details{
\tabular{ll}{
\strong{Application name} \tab  \strong{Keyword} \cr
Matrix Seriation \tab \code{seriate} \cr
}
}
\examples{
\dontrun{
deploy("seriate")
}
}
\seealso{
Other shiny: 
\code{\link{run_app}()}
}
\author{
N. Frerebeau
}
\concept{shiny}
