% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_diversity_alpha.R
\name{diversity_alpha_server}
\alias{diversity_alpha_server}
\title{Alpha Diversity Server}
\usage{
diversity_alpha_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{data.frame} (typically returned by \code{\link[=import_server]{import_server()}}).}
}
\description{
Alpha Diversity Server
}
\seealso{
\code{\link[=diversity_alpha_ui]{diversity_alpha_ui()}}

Other count data modules: 
\code{\link{diversity_alpha_ui}()},
\code{\link{diversity_beta_server}()},
\code{\link{diversity_beta_ui}()},
\code{\link{diversity_docs_ui}()},
\code{\link{occurrence_server}()},
\code{\link{occurrence_ui}()}
}
\concept{count data modules}
\keyword{internal}
