% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_diversity_beta.R
\name{diversity_beta_server}
\alias{diversity_beta_server}
\title{Beta Diversity Server}
\usage{
diversity_beta_server(id, x, y)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{data.frame} (typically returned by \code{\link[=import_server]{import_server()}}).}

\item{y}{A reactive \code{data.frame} returned by \code{\link[=diversity_alpha_server]{diversity_alpha_server()}}.}
}
\value{
A reactive \code{\link{data.frame}}.
}
\description{
Beta Diversity Server
}
\seealso{
\code{\link[=diversity_beta_ui]{diversity_beta_ui()}}

Other count data modules: 
\code{\link{count_server}()},
\code{\link{diversity_alpha_server}()},
\code{\link{diversity_alpha_ui}()},
\code{\link{diversity_beta_ui}()},
\code{\link{occurrence_server}()},
\code{\link{occurrence_ui}()}
}
\concept{count data modules}
\keyword{internal}
