% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_diversity_beta.R
\name{diversity_beta_ui}
\alias{diversity_beta_ui}
\title{Beta Diversity UI}
\usage{
diversity_beta_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Beta Diversity UI
}
\seealso{
\code{\link[=diversity_beta_server]{diversity_beta_server()}}

Other count data modules: 
\code{\link{diversity_alpha_server}()},
\code{\link{diversity_alpha_ui}()},
\code{\link{diversity_beta_server}()},
\code{\link{diversity_docs_ui}()},
\code{\link{occurrence_server}()},
\code{\link{occurrence_ui}()}
}
\concept{count data modules}
\keyword{internal}
