% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{diversity_server}
\alias{diversity_server}
\title{Alpha Diversity Server}
\usage{
diversity_server(id, x, verbose = get_option("verbose", FALSE))
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{data.frame} (typically returned by \code{\link[=import_server]{import_server()}}).}

\item{verbose}{A \code{\link{logical}} scalar: should \R report extra information on
progress?}
}
\value{
A reactive \code{\link{data.frame}} (see \code{\link[tabula:diversity]{tabula::diversity()}}).
}
\description{
Alpha Diversity Server
}
\seealso{
\code{\link[=diversity_ui]{diversity_ui()}}

Other count data modules: 
\code{\link{bertin_server}()},
\code{\link{bertin_ui}()},
\code{\link{count_server}()},
\code{\link{count_ui}()},
\code{\link{diversity_alpha_server}()},
\code{\link{diversity_alpha_ui}()},
\code{\link{diversity_beta_server}()},
\code{\link{diversity_beta_ui}()},
\code{\link{diversity_docs_ui}()},
\code{\link{diversity_ui}()},
\code{\link{occurrence_server}()},
\code{\link{occurrence_ui}()}
}
\concept{count data modules}
\keyword{internal}
