% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{export_multiple}
\alias{export_multiple}
\title{Download Multiple CSV Files}
\usage{
export_multiple(id, ..., name = "archive", label = tr_("Download results"))
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{...}{Further named arguments (\code{\link{data.frame}} to be saved).}

\item{name}{A \code{\link{character}} string specifying the name of the file
(without extension and the leading dot).}

\item{label}{A \code{\link{character}} string giving the label that should appear on
the button.}
}
\value{
No return value, called for side effects.
}
\description{
Save and Download several \code{\link{data.frame}} (Zip).
}
\keyword{internal}
