% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis-internal.R
\name{falsy}
\alias{falsy}
\alias{\%|||\%}
\title{Default Value for Falsy}
\usage{
x \%|||\% y
}
\arguments{
\item{x, y}{An object.}
}
\value{
If \code{x} is not \link[shiny:isTruthy]{truthy}, returns \code{y};
otherwise returns \code{x}.
}
\description{
Replaces a \link[shiny:isTruthy]{falsy} value with a default value.
}
\keyword{internal}
