% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home.R
\name{footer_server}
\alias{footer_server}
\title{Footer Server}
\usage{
footer_server(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
No return value, called for side effects.
}
\description{
Footer Server
}
\seealso{
\code{\link[=footer_ui]{footer_ui()}}

Other page modules: 
\code{\link{footer_ui}()},
\code{\link{home_server}()},
\code{\link{home_ui}()}
}
\concept{page modules}
\keyword{internal}
