% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{get_config}
\alias{get_config}
\title{Read Configuration Values}
\usage{
get_config(app = c("seriation"), config = getOption("janus.config"))
}
\arguments{
\item{app}{A \code{\link{character}} string specifying the Shiny application
configuration to get . Any unambiguous substring can be given.}

\item{config}{A \code{\link{character}} string specifying the name of configuration to
read from.}
}
\value{
A \code{\link{list}} of configuration values.
}
\description{
A wrapper for \code{\link[config:get]{config::get()}}.
}
\author{
N. Frerebeau
}
\keyword{internal}
