% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{graphics_server}
\alias{graphics_server}
\title{Graphical Parameters Server}
\usage{
graphics_server(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
A \code{\link[shiny:reactiveValues]{reactiveValues}} object with elements:
\describe{
\item{\code{col_quali}, \code{col_quant}}{A palette function that when called with a
single argument returns a character vector of colors.}
\item{\code{pch}}{A palette function that when called with a single argument
returns a character vector of symbols.}
\item{\code{lty}}{A palette function that when called with a single argument
returns a character vector of symbols.}
\item{\code{cex}}{A palette function that when called with a single argument
returns a numeric vector giving the amount by which plotting text and
symbols should be magnified relative to the default.}
\item{\code{asp}}{}
}
}
\description{
Graphical Parameters Server
}
\seealso{
\code{\link[=graphics_ui]{graphics_ui()}}

Other plot modules: 
\code{\link{graphics_ui}()},
\code{\link{scatter_server}()},
\code{\link{scatter_ui}()},
\code{\link{ternary_server}()},
\code{\link{ternary_ui}()}
}
\concept{plot modules}
\keyword{internal}
