% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{graphics_ui}
\alias{graphics_ui}
\title{Graphical Parameters UI}
\usage{
graphics_ui(
  id,
  col_quali = TRUE,
  col_quant = TRUE,
  pch = TRUE,
  lty = TRUE,
  cex = TRUE,
  asp = FALSE
)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\value{
A \code{\link{list}} of UI elements.
}
\description{
Graphical Parameters UI
}
\seealso{
\code{\link[=graphics_server]{graphics_server()}}

Other plot modules: 
\code{\link{graphics_server}()},
\code{\link{scatter_server}()},
\code{\link{scatter_ui}()},
\code{\link{ternary_server}()},
\code{\link{ternary_ui}()}
}
\concept{plot modules}
\keyword{internal}
