% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_home.R
\name{header_server}
\alias{header_server}
\title{Header Server}
\usage{
header_server(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\description{
Header Server
}
\seealso{
\code{\link[=header_ui]{header_ui()}}

Other page modules: 
\code{\link{footer_server}()},
\code{\link{footer_ui}()},
\code{\link{header_ui}()},
\code{\link{home_server}()},
\code{\link{home_ui}()}
}
\concept{page modules}
\keyword{internal}
