% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_home.R
\name{header_ui}
\alias{header_ui}
\title{Header UI}
\usage{
header_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Header UI
}
\seealso{
\code{\link[=header_server]{header_server()}}

Other page modules: 
\code{\link{footer_server}()},
\code{\link{footer_ui}()},
\code{\link{header_server}()},
\code{\link{home_server}()},
\code{\link{home_ui}()}
}
\concept{page modules}
\keyword{internal}
