% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home.R
\name{home_server}
\alias{home_server}
\title{Home Server}
\usage{
home_server(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
No return value, called for side effects.
}
\description{
Home Server
}
\seealso{
\code{\link[=home_ui]{home_ui()}}

Other page modules: 
\code{\link{footer_server}()},
\code{\link{footer_ui}()},
\code{\link{home_ui}()}
}
\concept{page modules}
\keyword{internal}
