% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_home.R
\name{home_ui}
\alias{home_ui}
\title{About UI}
\usage{
home_ui(id, name = NULL)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
About UI
}
\seealso{
\code{\link[=home_server]{home_server()}}

Other page modules: 
\code{\link{footer_server}()},
\code{\link{footer_ui}()},
\code{\link{header_server}()},
\code{\link{header_ui}()},
\code{\link{home_server}()}
}
\concept{page modules}
\keyword{internal}
