% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_ui}
\alias{import_ui}
\title{Import Data UI}
\usage{
import_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\value{
A \code{\link{list}} of UI elements.
}
\description{
Import Data UI
}
\seealso{
\code{\link[=import_server]{import_server()}}

Other generic modules: 
\code{\link{import_server}()},
\code{\link{prepare_server}()},
\code{\link{prepare_ui}()}
}
\concept{generic modules}
\keyword{internal}
