% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janus-package.R
\docType{package}
\name{janus-package}
\alias{janus-package}
\alias{_PACKAGE}
\alias{janus}
\title{janus: 'Shiny' Applications for the tesselle R Packages}
\description{
A collection of 'shiny' applications for the tesselle R
    packages.
}
\details{
\tabular{ll}{
\strong{Package:} \tab janus \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 0.0.0.9000 \cr
\strong{License:} \tab GPL-3 \cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tesselle/janus}
  \item Report bugs at \url{https://github.com/tesselle/janus/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

IRAMAT-CRP2A (UMR 5060)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
