% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin.batch.R
\name{kin.batch.process}
\alias{kin.batch.process}
\title{Process multiple trials in parallel}
\usage{
kin.batch.process(
  data,
  start_positions,
  end_positions,
  refresh_rate = 1/120,
  n_cores = NULL,
  progress = TRUE
)
}
\arguments{
\item{data}{Data frame containing all trials}

\item{start_positions}{List of start positions for each trial}

\item{end_positions}{List of end positions for each trial}

\item{refresh_rate}{Sampling rate in seconds}

\item{n_cores}{Number of cores to use for parallel processing}

\item{progress}{Show progress bar}
}
\value{
List of processed trials
}
\description{
Process multiple trials in parallel
}
