% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin.bwFilter.r
\name{kin.bwFilter}
\alias{kin.bwFilter}
\title{apply Butterworth filter}
\usage{
kin.bwFilter(s, n = 2, cutoff_freq = 10, type = "low")
}
\arguments{
\item{s:}{signal}

\item{n:}{filter order (see ?butter for details)}

\item{W:}{critical frequencies of the filter (see ?butter for details)}

\item{type:}{filter type (see ?butter for details)}
}
\description{
apply Butterworth filter
}
