% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin.extract.r
\name{kin.extract}
\alias{kin.extract}
\title{Extract various kinematics from a grasp}
\usage{
kin.extract(dataset, MGA.range = NULL, TGA.type = "s", TGA.lim = 20,
  TGA.kin = "g", graphs = F, ...)
}
\arguments{
\item{dataset}{an object of the type data.frame}

\item{MGA.range}{optional: a two elements vector with the interval (in frames) within which to look for the MGA (the greatest value of grip aperture within the interval). If not specified then the range is the entire grasp}

\item{TGA.type}{character: "s" tells the function to extract the TGA based on a spatial threshold (default); "v" extracts the TGA based on a velocity threshold. In both cases the threshold value is specified by the parameter TGA.lim}

\item{TGA.lim}{the value of the threshold for the extraction of the TGA}

\item{graphs}{logic: should a summary graph be printed? (default to false)}
}
\description{
Extract various kinematics from a grasp
}
\examples{
libraries()

data(rtgData)
rtgData <- data.check(rtgData)
testData <- kin.SmoothAndDerive(rtgData)
system.time(testData.kin <- kin.extract(testData))
system.time(testData.kin2 <- kin.extract(testData, MGA.range=c(70,90)))

view.summary.trials.GA(21:40, testData, testData.kin, print=F)
view.summary.trials.GA(21:40, testData, testData.kin2, print=F)

detach(debugData)
}
