% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dataset_fix.r
\name{kin.frameN}
\alias{kin.frameN}
\title{Create Frame Counter}
\usage{
kin.frameN(dataset)
}
\arguments{
\item{dataset}{A data frame containing kinematic data}
}
\value{
A data frame with added frame numbers
}
\description{
Creates sequential frame numbers within each trial.
}
