% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dataset_fix.r
\name{kin.globalTime}
\alias{kin.globalTime}
\title{Calculate Global Time}
\usage{
kin.globalTime(dataset)
}
\arguments{
\item{dataset}{A data frame containing kinematic data with time column}
}
\value{
A data frame with added globalTime column
}
\description{
Calculates global time across all trials based on median frame duration.
}
