% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{kin.max}
\alias{kin.max}
\title{Maximum Value with NA Handling}
\usage{
kin.max(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Maximum value in the vector
}
\description{
Wrapper for max() with na.rm=TRUE by default
}
