% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{kin.mean}
\alias{kin.mean}
\title{Mean with NA Handling}
\usage{
kin.mean(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Mean value of the vector
}
\description{
Wrapper for mean() with na.rm=TRUE by default
}
