% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{kin.min}
\alias{kin.min}
\title{Minimum Value with NA Handling}
\usage{
kin.min(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Minimum value in the vector
}
\description{
Wrapper for min() with na.rm=TRUE by default
}
