% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin.visualization.R
\name{kin.plot.grasp}
\alias{kin.plot.grasp}
\title{Plot grip aperture}
\usage{
kin.plot.grasp(data, time_col, aperture_col, title = "Grip Aperture")
}
\arguments{
\item{data}{Data frame containing grip aperture data}

\item{time_col}{Column name for time}

\item{aperture_col}{Column name for grip aperture}

\item{title}{Plot title}
}
\value{
A line plot using ggplot2
}
\description{
Plot grip aperture
}
