% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin.visualization.R
\name{kin.plot.trajectory}
\alias{kin.plot.trajectory}
\title{Plot 3D trajectory}
\usage{
kin.plot.trajectory(
  data,
  x_col,
  y_col,
  z_col,
  color_by = NULL,
  title = "3D Trajectory"
)
}
\arguments{
\item{data}{Data frame containing trajectory data}

\item{x_col}{Column name for x coordinates}

\item{y_col}{Column name for y coordinates}

\item{z_col}{Column name for z coordinates}

\item{color_by}{Optional column to color points by}

\item{title}{Plot title}
}
\value{
A 3D plot using plotly
}
\description{
Plot 3D trajectory
}
