% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin.visualization.R
\name{kin.plot.velocity}
\alias{kin.plot.velocity}
\title{Plot velocity profile}
\usage{
kin.plot.velocity(data, time_col, velocity_col, title = "Velocity Profile")
}
\arguments{
\item{data}{Data frame containing velocity data}

\item{time_col}{Column name for time}

\item{velocity_col}{Column name for velocity}

\item{title}{Plot title}
}
\value{
A line plot using ggplot2
}
\description{
Plot velocity profile
}
