% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_smooth_derive.r
\name{kin.rescale}
\alias{kin.rescale}
\title{rescale the values of a vector within a specified range. Inputs:}
\usage{
kin.rescale(x, a, b)
}
\arguments{
\item{x:}{vector to be rescaled}

\item{a:}{range minimum}

\item{b:}{range maximum}
}
\description{
rescale the values of a vector within a specified range. Inputs:
}
