% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin.rotate.trajectory.r
\name{kin.rotate.trajectory}
\alias{kin.rotate.trajectory}
\title{Rotate trajectory}
\usage{
kin.rotate.trajectory(data, end, f = T, t = T, s = T)
}
\arguments{
\item{data}{a rX3 matrix of x,y,z observations (data.frame format is automatically converted to matrix)}

\item{end}{a 3-length vector with x,y,z coordinates of the trajectory endpoint}
}
\description{
Rotate trajectory
}
\details{
The trajectory is assumed to start at origin (0,0,0)
and terminate at a point (0,0,z)
with x axis: rightwards positive, leftwards negative
y axis: upwards positive, downwards negative
z axis: forwards positive, backwards negative
}
