% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{kin.se}
\alias{kin.se}
\title{Standard Error}
\usage{
kin.se(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{Logical; should NA values be removed?}
}
\value{
Standard error value
}
\description{
Calculates the standard error of a vector
}
