% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin.signal.analysis.R
\name{kin.signal.analysis}
\alias{kin.signal.analysis}
\title{Analyze trajectory data}
\usage{
kin.signal.analysis(
  signal,
  signal.name = "signal",
  start,
  end,
  maxFrames = 20,
  splinepar = 0.5,
  rotate = T,
  f = T,
  t = T,
  s = T,
  ...
)
}
\arguments{
\item{signal}{dataset to be analyzed. It must contain a time column and three columns for the x, y, z positions of the object (signal) that will be analyzed. See details for the expected names of these columns}

\item{signal.name}{character indicating the name of the object (signal)}

\item{start}{3-elements-long vector indicating the x, y, z coordinates of the start position of the trajectory}

\item{end}{3-elements-long vector indicating the x, y, z coordinates of the end position of the trajectory}

\item{maxFrames}{integer number of missing frames to be substituted with linear interpolation}

\item{splinepar}{parameter for spline smoothing (see ?smooth.spline)}

\item{rotate}{logical: should the trajectory be rotated to align to the start-to-end direction?}

\item{f}{logical (requires rotate = TRUE): should the trajectory be aligned to the frontoparallel plane?}

\item{t}{logical (requires rotate = TRUE): should the trajectory be aligned to the transversal plane?}

\item{s}{logical (requires rotate = TRUE): should the trajectory be aligned to the sagittal plane?}
}
\description{
Repair (check for missing frames), smooth (spline), derive, rotate trajectory data
}
\details{
The names of the input dataset (signal parameter) must adhere to the following standards:
i) the name of the time column must be identical to what specified through kin.setDataCols()
ii) the other three columns names must be string of the type nameXraw, nameYraw, nameZraw (where "name" must match the signal.name parameter)
}
