% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin.signal.arclength.R
\name{kin.signal.arclength}
\alias{kin.signal.arclength}
\title{Calculate Cumulative Arc Length of 3D Motion Data}
\usage{
kin.signal.arclength(data)
}
\arguments{
\item{data}{A matrix or data frame with 3 columns representing X, Y, and Z coordinates}
}
\value{
A numeric vector containing the cumulative arc length at each point
}
\description{
This function calculates the cumulative arc length of a 3D motion trajectory
using Savitzky-Golay filtered derivatives.
}
\details{
The function computes the arc length by:
\enumerate{
\item Calculating derivatives using Savitzky-Golay filtering for each dimension
\item Computing the Euclidean distance between consecutive points
\item Calculating the cumulative sum of these distances
}
}
\examples{
xyz_data <- matrix(rnorm(300), ncol=3)
arc_length <- kin.signal.arclength(xyz_data)

}
