% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dataset_fix.r
\name{kin.signal.missing}
\alias{kin.signal.missing}
\title{Identify Missing Signals}
\usage{
kin.signal.missing(x, criterion = NULL, delete.static.positions = F)
}
\arguments{
\item{x}{A numeric vector containing signal data}

\item{criterion}{Values to be considered as missing (default: NULL)}

\item{delete.static.positions}{If TRUE, marks unchanging positions as missing (default: FALSE)}
}
\value{
A vector with missing values marked as NA
}
\description{
Identifies missing or static signals in kinematic data.
}
