% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin.SmoothAndDerive.r
\name{kin.SmoothAndDerive}
\alias{kin.SmoothAndDerive}
\title{Smooth and derive kinematics}
\usage{
kin.SmoothAndDerive(dataset, lambda = NULL, frames.to.interpolate = 18,
  index = T, thumb = T, reverse.Zcoords = T, ...)
}
\arguments{
\item{dataset}{an object of the type data.frame}

\item{lambda}{smoothing factor. If not specified, it is = 5e-04}

\item{frames.to.interpolate}{the minimum amount of frames across which to interpolate during the spline fit in case of missing data. Default to 18}

\item{index}{logical: is the index included in the dataset?}

\item{thumb}{logical: is the thumb included in the dataset?}

\item{reverse.Zcoords}{logical: should the Z coordinates be multiplied by -1?}
}
\description{
Smooth and derive kinematics
}
\examples{
libraries()

data(rtgData)
rtgData <- data.check(rtgData)
system.time(testData <- kin.SmoothAndDerive(rtgData))

view.summary.trials(21:40, testData, print=F)

detach(rtgData)
}
