% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_dataset_fix.r
\name{kin.trialN}
\alias{kin.trialN}
\title{Fix Trial Numbers}
\usage{
kin.trialN(dataset)
}
\arguments{
\item{dataset}{A data frame containing kinematic data}
}
\value{
A data frame with corrected trial numbers
}
\description{
Ensures trial numbers start from 1 by adding 1 to all trial numbers if any are 0.
}
