% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{lengthunique}
\alias{lengthunique}
\title{Length of Unique Elements}
\usage{
lengthunique(x)
}
\arguments{
\item{x}{A vector}
}
\value{
Integer representing count of unique elements
}
\description{
Returns the number of unique elements in a vector
}
