% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libraries.r
\name{libraries}
\alias{libraries}
\title{Load and update all the required libraries of the shapelab package}
\usage{
libraries(update = F)
}
\arguments{
\item{update}{if TRUE, the function also updates the libraries to newest version (FALSE by default)}
}
\description{
Load and update all the required libraries of the shapelab package
}
