% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{lm_server}
\alias{lm_server}
\title{Linear Model Server}
\usage{
lm_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{\link{data.frame}}.}
}
\value{
A reactive \code{\link{lm}} object.
}
\description{
Linear Model Server
}
\seealso{
\code{\link[=lm_ui]{lm_ui()}}

Other modeling modules: 
\code{\link{lm_ui}()}
}
\concept{modeling modules}
\keyword{internal}
