% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{lm_ui}
\alias{lm_ui}
\title{Linear Model UI}
\usage{
lm_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\value{
A nav item that may be passed to a nav container
(e.g. \code{\link[bslib:navset]{bslib::navset_tab()}}).
}
\description{
Linear Model UI
}
\seealso{
\code{\link[=lm_server]{lm_server()}}

Other modeling modules: 
\code{\link{lm_server}()}
}
\concept{modeling modules}
\keyword{internal}
