% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{lmax.locate}
\alias{lmax.locate}
\title{Locate Local Maxima}
\usage{
lmax.locate(x, y)
}
\arguments{
\item{x}{Numeric vector of x coordinates}

\item{y}{Numeric vector of y coordinates}
}
\value{
Logical vector indicating which points are local maxima
}
\description{
Identifies points that are local maxima in a curve
}
