% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_logratio.R
\name{logratio_server}
\alias{logratio_server}
\title{Log-Ratio Server}
\usage{
logratio_server(id, x, method)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{\link[nexus:CompositionMatrix-class]{nexus::CompositionMatrix}} object.}

\item{method}{A \code{\link{character}} string specifying the log-ratio to be computed.}
}
\value{
A reactive \code{\link[nexus:LogRatio-class]{nexus::LogRatio}} object.
}
\description{
Log-Ratio Server
}
\seealso{
\code{\link[=logratio_ui]{logratio_ui()}}

Other coda modules: 
\code{\link{coda_barplot_server}()},
\code{\link{coda_barplot_ui}()},
\code{\link{coda_hclust_server}()},
\code{\link{coda_hclust_ui}()},
\code{\link{coda_outliers_server}()},
\code{\link{coda_outliers_ui}()},
\code{\link{coda_server}()},
\code{\link{coda_summary_server}()},
\code{\link{coda_summary_ui}()},
\code{\link{coda_ui}()},
\code{\link{logratio_ui}()}
}
\concept{coda modules}
\keyword{internal}
