% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_logratio.R
\name{logratio_ui}
\alias{logratio_ui}
\title{Log-Ratio UI}
\usage{
logratio_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Log-Ratio UI
}
\seealso{
\code{\link[=logratio_server]{logratio_server()}}

Other coda modules: 
\code{\link{coda_outliers_server}()},
\code{\link{coda_outliers_ui}()},
\code{\link{coda_plot_server}()},
\code{\link{coda_plot_ui}()},
\code{\link{coda_server}()},
\code{\link{coda_summary_server}()},
\code{\link{coda_summary_ui}()},
\code{\link{coda_ui}()},
\code{\link{logratio_server}()}
}
\concept{coda modules}
\keyword{internal}
