% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{maxima}
\alias{maxima}
\title{Find Local Maxima}
\usage{
maxima(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Vector of indices where local maxima occur
}
\description{
Identifies indices of local maxima in a vector
}
