% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_mcd.R
\name{mcd_server}
\alias{mcd_server}
\title{MCD Server}
\usage{
mcd_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{data.frame} (typically returned by \code{\link[=import_server]{import_server()}}).}
}
\value{
No return value, called for side effects.
}
\description{
MCD Server
}
\seealso{
\code{\link[=mcd_ui]{mcd_ui()}}

Other ceramic data modules: 
\code{\link{mcd_ui}()}
}
\concept{ceramic data modules}
\keyword{internal}
