% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{minima}
\alias{minima}
\title{Find Local Minima}
\usage{
minima(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Vector of indices where local minima occur
}
\description{
Identifies indices of local minima in a vector
}
