% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data.R
\name{missing_ui}
\alias{missing_ui}
\title{Missing Values UI}
\usage{
missing_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Missing Values UI
}
\seealso{
\code{\link[=missing_server]{missing_server()}}

Other generic modules: 
\code{\link{data_server}()},
\code{\link{data_ui}()},
\code{\link{import_server}()},
\code{\link{import_ui}()},
\code{\link{missing_server}()},
\code{\link{prepare_server}()},
\code{\link{prepare_ui}()}
}
\concept{generic modules}
\keyword{internal}
