% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_about.R
\name{module_about_ui}
\alias{module_about_ui}
\title{About UI}
\usage{
module_about_ui(id, cite = "janus")
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
About UI
}
\seealso{
Other UI modules: 
\code{\link{module_ca_ui}()},
\code{\link{module_footer_ui}()},
\code{\link{module_import_ui}()},
\code{\link{module_seriate_ui}()},
\code{\link{module_settings_ui}()}
}
\concept{UI modules}
\keyword{internal}
