% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_ca_ui.R
\name{module_ca_ui}
\alias{module_ca_ui}
\title{Correspondence Analysis UI}
\usage{
module_ca_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Correspondence Analysis UI
}
\seealso{
\code{\link[=module_ca_server]{module_ca_server()}}

Other UI modules: 
\code{\link{module_about_ui}()},
\code{\link{module_footer_ui}()},
\code{\link{module_import_ui}()},
\code{\link{module_seriate_ui}()},
\code{\link{module_visualize_ui}()}
}
\concept{UI modules}
