% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_chi2.R
\name{module_chi_server}
\alias{module_chi_server}
\title{Chi-squared Test Server}
\usage{
module_chi_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{data.frame} (typically returned by
\code{\link[=module_prepare_server]{module_prepare_server()}}).}
}
\description{
Chi-squared Test Server
}
\seealso{
\code{\link[=module_chi2_ui]{module_chi2_ui()}}

Other server modules: 
\code{\link{module_count_server}()},
\code{\link{module_footer_server}()},
\code{\link{module_header_server}()},
\code{\link{module_home_server}()},
\code{\link{module_import_server}()},
\code{\link{module_missing_server}()},
\code{\link{module_multivar_server}()},
\code{\link{module_prepare_server}()},
\code{\link{module_seriate_server}()},
\code{\link{module_summary_server}()}
}
\concept{server modules}
\keyword{internal}
