% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_coda_plot.R
\name{module_coda_plot_server}
\alias{module_coda_plot_server}
\title{Bar Plot Server}
\usage{
module_coda_plot_server(id, user_data)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{user_data}{A reactive \code{\link[nexus:CompositionMatrix-class]{nexus::CompositionMatrix}} object.}
}
\description{
Bar Plot Server
}
\seealso{
\code{\link[=module_coda_plot_ui]{module_coda_plot_ui()}}

Other coda modules: 
\code{\link{module_coda_plot_ui}()},
\code{\link{module_coda_server}()},
\code{\link{module_coda_summary_server}()},
\code{\link{module_coda_summary_ui}()},
\code{\link{module_coda_ui}()},
\code{\link{module_logratio_server}()},
\code{\link{module_logratio_ui}()}
}
\concept{coda modules}
\keyword{internal}
