% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_coda.R
\name{module_coda_server}
\alias{module_coda_server}
\title{Compositional Data Server}
\usage{
module_coda_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{data.frame} (typically returned by
\code{\link[=module_import_server]{module_import_server()}}).}
}
\value{
A reactive \code{\link[nexus:CompositionMatrix-class]{nexus::CompositionMatrix}} object.
}
\description{
Compositional Data Server
}
\seealso{
\code{\link[=module_coda_ui]{module_coda_ui()}}

Other coda modules: 
\code{\link{module_coda_plot_server}()},
\code{\link{module_coda_plot_ui}()},
\code{\link{module_coda_summary_server}()},
\code{\link{module_coda_summary_ui}()},
\code{\link{module_coda_ui}()},
\code{\link{module_logratio_server}()},
\code{\link{module_logratio_ui}()}
}
\concept{coda modules}
\keyword{internal}
