% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_coda_summary.R
\name{module_coda_summary_server}
\alias{module_coda_summary_server}
\title{Compositional Data Summary Server}
\usage{
module_coda_summary_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{\link[nexus:CompositionMatrix-class]{nexus::CompositionMatrix}} object.}
}
\description{
Compositional Data Summary Server
}
\seealso{
\code{\link[=module_coda_summary_ui]{module_coda_summary_ui()}}

Other coda modules: 
\code{\link{module_coda_plot_server}()},
\code{\link{module_coda_plot_ui}()},
\code{\link{module_coda_server}()},
\code{\link{module_coda_summary_ui}()},
\code{\link{module_coda_ui}()},
\code{\link{module_logratio_server}()},
\code{\link{module_logratio_ui}()}
}
\concept{coda modules}
\keyword{internal}
