% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_coda_summary.R
\name{module_coda_summary_ui}
\alias{module_coda_summary_ui}
\title{Compositional Data Summary UI}
\usage{
module_coda_summary_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Compositional Data Summary UI
}
\seealso{
\code{\link[=module_coda_summary_server]{module_coda_summary_server()}}

Other coda modules: 
\code{\link{module_coda_plot_server}()},
\code{\link{module_coda_plot_ui}()},
\code{\link{module_coda_server}()},
\code{\link{module_coda_summary_server}()},
\code{\link{module_coda_ui}()},
\code{\link{module_logratio_server}()},
\code{\link{module_logratio_ui}()}
}
\concept{coda modules}
\keyword{internal}
