% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data.R
\name{module_data_ui}
\alias{module_data_ui}
\title{Data UI}
\usage{
module_data_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Data UI
}
\seealso{
Other UI modules: 
\code{\link{module_about_ui}()},
\code{\link{module_ca_ui}()},
\code{\link{module_footer_ui}()},
\code{\link{module_import_ui}()},
\code{\link{module_prepare_ui}()},
\code{\link{module_seriate_ui}()},
\code{\link{module_settings_ui}()}
}
\concept{UI modules}
\keyword{internal}
