% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_footer.R
\name{module_footer_ui}
\alias{module_footer_ui}
\title{Footer UI}
\usage{
module_footer_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}

\item{package}{A \code{\link{character}} string giving the name of the app.}
}
\description{
Footer UI
}
\seealso{
\code{\link[=module_footer_server]{module_footer_server()}}

Other UI modules: 
\code{\link{module_chi2_ui}()},
\code{\link{module_header_ui}()},
\code{\link{module_home_ui}()},
\code{\link{module_import_ui}()},
\code{\link{module_missing_ui}()},
\code{\link{module_multivar_ui}()},
\code{\link{module_prepare_ui}()},
\code{\link{module_seriate_ui}()},
\code{\link{module_summary_ui}()}
}
\concept{UI modules}
\keyword{internal}
