% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_header.R
\name{module_header_ui}
\alias{module_header_ui}
\title{Header UI}
\usage{
module_header_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Header UI
}
\seealso{
\code{\link[=module_header_server]{module_header_server()}}

Other UI modules: 
\code{\link{module_footer_ui}()},
\code{\link{module_home_ui}()},
\code{\link{module_import_ui}()},
\code{\link{module_missing_ui}()},
\code{\link{module_multivar_ui}()},
\code{\link{module_prepare_ui}()},
\code{\link{module_seriate_ui}()},
\code{\link{module_summary_ui}()}
}
\concept{UI modules}
\keyword{internal}
