% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_home.R
\name{module_home_server}
\alias{module_home_server}
\title{Home Server}
\usage{
module_home_server(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\description{
Home Server
}
\seealso{
\code{\link[=module_home_ui]{module_home_ui()}}

Other server modules: 
\code{\link{module_count_server}()},
\code{\link{module_footer_server}()},
\code{\link{module_header_server}()},
\code{\link{module_import_server}()},
\code{\link{module_missing_server}()},
\code{\link{module_multivar_server}()},
\code{\link{module_prepare_server}()},
\code{\link{module_seriate_server}()},
\code{\link{module_summary_server}()}
}
\concept{server modules}
\keyword{internal}
