% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_import_ui.R
\name{module_import_ui}
\alias{module_import_ui}
\title{Import UI}
\usage{
module_import_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Import UI
}
\seealso{
\code{\link[=module_import_server]{module_import_server()}}

Other UI modules: 
\code{\link{module_about_ui}()},
\code{\link{module_ca_ui}()},
\code{\link{module_footer_ui}()},
\code{\link{module_seriate_ui}()},
\code{\link{module_settings_ui}()}
}
\concept{UI modules}
