% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_coda_logratio.R
\name{module_logratio_server}
\alias{module_logratio_server}
\title{Log-Ratio Server}
\usage{
module_logratio_server(id, x, method)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{\link[nexus:CompositionMatrix-class]{nexus::CompositionMatrix}} object.}

\item{method}{A \code{\link{character}} string specifying the log-ratio to be computed.}
}
\value{
A reactive \code{\link[nexus:LogRatio-class]{nexus::LogRatio}} object.
}
\description{
Log-Ratio Server
}
\seealso{
\code{\link[=module_logratio_ui]{module_logratio_ui()}}

Other coda modules: 
\code{\link{module_coda_plot_server}()},
\code{\link{module_coda_plot_ui}()},
\code{\link{module_coda_server}()},
\code{\link{module_coda_summary_server}()},
\code{\link{module_coda_summary_ui}()},
\code{\link{module_coda_ui}()},
\code{\link{module_logratio_ui}()}
}
\concept{coda modules}
\keyword{internal}
