% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_coda_logratio.R
\name{module_logratio_ui}
\alias{module_logratio_ui}
\title{Log-Ratio UI}
\usage{
module_logratio_ui()
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Log-Ratio UI
}
\seealso{
\code{\link[=module_logratio_server]{module_logratio_server()}}

Other coda modules: 
\code{\link{module_coda_plot_server}()},
\code{\link{module_coda_plot_ui}()},
\code{\link{module_coda_server}()},
\code{\link{module_coda_summary_server}()},
\code{\link{module_coda_summary_ui}()},
\code{\link{module_coda_ui}()},
\code{\link{module_logratio_server}()}
}
\concept{coda modules}
\keyword{internal}
