% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_multivar.R
\name{module_multivar_server}
\alias{module_multivar_server}
\title{Multivariate Analysis Server}
\usage{
module_multivar_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{\link[dimensio:MultivariateAnalysis]{dimensio::MultivariateAnalysis}} object.}
}
\description{
Multivariate Analysis Server
}
\seealso{
\link{module_multivar_ui}

Other server modules: 
\code{\link{module_chi_server}()},
\code{\link{module_count_server}()},
\code{\link{module_footer_server}()},
\code{\link{module_header_server}()},
\code{\link{module_home_server}()},
\code{\link{module_import_server}()},
\code{\link{module_missing_server}()},
\code{\link{module_pca_server}()},
\code{\link{module_prepare_server}()},
\code{\link{module_seriate_server}()},
\code{\link{module_summary_server}()},
\code{\link{module_ternary_server}()}
}
\concept{server modules}
\keyword{internal}
