% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_multivar.R
\name{module_multivar_ui}
\alias{module_multivar_ui}
\title{Mutlivariate Analysis UI}
\usage{
module_multivar_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\value{
A tab that can be passed to \code{\link[shiny:tabsetPanel]{shiny::tabsetPanel()}}.
}
\description{
Mutlivariate Analysis UI
}
\seealso{
\code{\link[=module_multivar_server]{module_multivar_server()}}

Other UI modules: 
\code{\link{module_footer_ui}()},
\code{\link{module_header_ui}()},
\code{\link{module_home_ui}()},
\code{\link{module_import_ui}()},
\code{\link{module_missing_ui}()},
\code{\link{module_prepare_ui}()},
\code{\link{module_seriate_ui}()},
\code{\link{module_summary_ui}()}
}
\concept{UI modules}
\keyword{internal}
