% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_pca.R
\name{module_pca_server}
\alias{module_pca_server}
\title{Principal Components Analysis Server}
\usage{
module_pca_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A list of reactive \code{data.frame}.}
}
\value{
A reactive \code{\link[dimensio:PCA-class]{dimensio::PCA}} object.
}
\description{
Principal Components Analysis Server
}
\seealso{
\code{\link[=module_pca_ui]{module_pca_ui()}}

Other server modules: 
\code{\link{module_chi_server}()},
\code{\link{module_count_server}()},
\code{\link{module_footer_server}()},
\code{\link{module_header_server}()},
\code{\link{module_home_server}()},
\code{\link{module_import_server}()},
\code{\link{module_missing_server}()},
\code{\link{module_multivar_server}()},
\code{\link{module_prepare_server}()},
\code{\link{module_seriate_server}()},
\code{\link{module_summary_server}()},
\code{\link{module_ternary_server}()}
}
\concept{server modules}
\keyword{internal}
