% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_pca.R
\name{module_pca_ui}
\alias{module_pca_ui}
\title{Principal Components Analysis UI}
\usage{
module_pca_ui(id, center = TRUE, scale = TRUE)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}

\item{center}{A \code{\link{logical}} scalar: should the variables be shifted to be
zero centered?}

\item{scale}{A \code{\link{logical}} scalar: should the variables be scaled to unit
variance?}
}
\description{
Principal Components Analysis UI
}
\seealso{
\code{\link[=module_pca_server]{module_pca_server()}}

Other UI modules: 
\code{\link{module_chi2_ui}()},
\code{\link{module_footer_ui}()},
\code{\link{module_header_ui}()},
\code{\link{module_home_ui}()},
\code{\link{module_import_ui}()},
\code{\link{module_missing_ui}()},
\code{\link{module_multivar_ui}()},
\code{\link{module_prepare_ui}()},
\code{\link{module_seriate_ui}()},
\code{\link{module_summary_ui}()},
\code{\link{module_ternary_ui}()}
}
\concept{UI modules}
\keyword{internal}
