% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_prepare.R
\name{module_prepare_server}
\alias{module_prepare_server}
\title{Prepare Data Server}
\usage{
module_prepare_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{data.frame} (typically returned by
\code{\link[=module_import_server]{module_import_server()}}).}
}
\value{
A reactive \code{data.frame}.
}
\description{
Prepare Data Server
}
\seealso{
\code{\link[=module_prepare_ui]{module_prepare_ui()}}

Other server modules: 
\code{\link{module_count_server}()},
\code{\link{module_footer_server}()},
\code{\link{module_header_server}()},
\code{\link{module_home_server}()},
\code{\link{module_import_server}()},
\code{\link{module_missing_server}()},
\code{\link{module_multivar_server}()},
\code{\link{module_seriate_server}()},
\code{\link{module_summary_server}()}
}
\concept{server modules}
\keyword{internal}
