% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_prepare.R
\name{module_prepare_ui}
\alias{module_prepare_ui}
\title{Prepare Data UI}
\usage{
module_prepare_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Prepare Data UI
}
\seealso{
\code{\link[=module_prepare_server]{module_prepare_server()}}

Other UI modules: 
\code{\link{module_chi2_ui}()},
\code{\link{module_footer_ui}()},
\code{\link{module_header_ui}()},
\code{\link{module_home_ui}()},
\code{\link{module_import_ui}()},
\code{\link{module_missing_ui}()},
\code{\link{module_multivar_ui}()},
\code{\link{module_seriate_ui}()},
\code{\link{module_summary_ui}()}
}
\concept{UI modules}
\keyword{internal}
