% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_seriate.R
\name{module_seriate_server}
\alias{module_seriate_server}
\title{Seriate Server}
\usage{
module_seriate_server(id, user_data, user_settings)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{user_data}{A \code{\link[shiny:reactiveValues]{shiny::reactiveValues()}} list with the
following elements: "\code{data}".}

\item{user_settings}{A \code{\link[shiny:reactiveValues]{shiny::reactiveValues()}} list.}
}
\description{
Seriate Server
}
\seealso{
\code{\link[=module_seriate_ui]{module_seriate_ui()}}

Other server modules: 
\code{\link{module_ca_server}()},
\code{\link{module_footer_server}()},
\code{\link{module_import_server}()},
\code{\link{module_settings_server}()}
}
\concept{server modules}
\keyword{internal}
