% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_seriate.R
\name{module_seriate_ui}
\alias{module_seriate_ui}
\title{Seriate UI}
\usage{
module_seriate_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Seriate UI
}
\seealso{
\code{\link[=module_seriate_server]{module_seriate_server()}}

Other UI modules: 
\code{\link{module_about_ui}()},
\code{\link{module_ca_ui}()},
\code{\link{module_footer_ui}()},
\code{\link{module_import_ui}()},
\code{\link{module_settings_ui}()}
}
\concept{UI modules}
\keyword{internal}
