% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_settings.R
\name{module_settings_server}
\alias{module_settings_server}
\title{Settings Server}
\usage{
module_settings_server(id, user_settings)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{user_settings}{A \code{\link[shiny:reactiveValues]{shiny::reactiveValues()}} list.}
}
\description{
Settings Server
}
\seealso{
\code{\link[=module_settings_ui]{module_settings_ui()}}

Other server modules: 
\code{\link{module_ca_server}()},
\code{\link{module_footer_server}()},
\code{\link{module_import_server}()},
\code{\link{module_seriate_server}()}
}
\concept{server modules}
\keyword{internal}
