% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_settings.R
\name{module_settings_ui}
\alias{module_settings_ui}
\title{Settings UI}
\usage{
module_settings_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Settings UI
}
\seealso{
\code{\link[=module_settings_server]{module_settings_server()}}

Other UI modules: 
\code{\link{module_about_ui}()},
\code{\link{module_ca_ui}()},
\code{\link{module_data_ui}()},
\code{\link{module_footer_ui}()},
\code{\link{module_import_ui}()},
\code{\link{module_prepare_ui}()},
\code{\link{module_seriate_ui}()}
}
\concept{UI modules}
\keyword{internal}
