% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_summary.R
\name{module_summary_server}
\alias{module_summary_server}
\title{Data Summary Server}
\usage{
module_summary_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{data.frame} (typically returned by
\code{\link[=module_prepare_server]{module_prepare_server()}}).}
}
\description{
Data Summary Server
}
\seealso{
\code{\link[=module_summary_ui]{module_summary_ui()}}

Other server modules: 
\code{\link{module_chi_server}()},
\code{\link{module_count_server}()},
\code{\link{module_footer_server}()},
\code{\link{module_header_server}()},
\code{\link{module_home_server}()},
\code{\link{module_import_server}()},
\code{\link{module_missing_server}()},
\code{\link{module_multivar_server}()},
\code{\link{module_pca_server}()},
\code{\link{module_prepare_server}()},
\code{\link{module_seriate_server}()},
\code{\link{module_ternary_server}()}
}
\concept{server modules}
\keyword{internal}
