% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_visualize_server.R
\name{module_visualize_server}
\alias{module_visualize_server}
\title{Visualize Server}
\usage{
module_visualize_server(input, output, session, user_data, user_settings)
}
\arguments{
\item{input}{Provided by \pkg{Shiny}.}

\item{output}{Provided by \pkg{Shiny}.}

\item{session}{Provided by \pkg{Shiny}.}

\item{user_data}{A \code{\link[shiny:reactiveValues]{shiny::reactiveValues()}} list with the
following elements: "\code{data}".}

\item{user_settings}{A \code{\link[shiny:reactiveValues]{shiny::reactiveValues()}} list.}
}
\description{
Visualize Server
}
\seealso{
\code{\link[=module_visualize_ui]{module_visualize_ui()}}

Other Server modules: 
\code{\link{module_footer_server}()},
\code{\link{module_import_server}()},
\code{\link{module_seriate_server}()}
}
\concept{Server modules}
