% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_visualize_ui.R
\name{module_visualize_ui}
\alias{module_visualize_ui}
\title{Visualize UI}
\usage{
module_visualize_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Visualize UI
}
\seealso{
\code{\link[=module_visualize_server]{module_visualize_server()}}

Other UI modules: 
\code{\link{module_about_ui}()},
\code{\link{module_footer_ui}()},
\code{\link{module_import_ui}()},
\code{\link{module_seriate_ui}()}
}
\concept{UI modules}
