% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{multivariate_server}
\alias{multivariate_server}
\title{Multivariate Analysis Server}
\usage{
multivariate_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{\link[dimensio:MultivariateAnalysis]{dimensio::MultivariateAnalysis}} object.}
}
\description{
Multivariate Analysis Server
}
\seealso{
\link{multivariate_ui}

Other multivariate analysis modules: 
\code{\link{ca_server}()},
\code{\link{ca_ui}()},
\code{\link{multivariate_ui}()},
\code{\link{pca_server}()},
\code{\link{pca_ui}()}
}
\concept{multivariate analysis modules}
\keyword{internal}
