% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_occurrence.R
\name{occurrence_server}
\alias{occurrence_server}
\title{Co-Occurrence Server}
\usage{
occurrence_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{data.frame} (typically returned by \code{\link[=import_server]{import_server()}}).}
}
\value{
A reactive \code{\link{data.frame}}.
}
\description{
Co-Occurrence Server
}
\seealso{
\code{\link[=occurrence_ui]{occurrence_ui()}}

Other count data modules: 
\code{\link{diversity_alpha_server}()},
\code{\link{diversity_alpha_ui}()},
\code{\link{diversity_beta_server}()},
\code{\link{diversity_beta_ui}()},
\code{\link{diversity_docs_ui}()},
\code{\link{occurrence_ui}()}
}
\concept{count data modules}
\keyword{internal}
