% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_occurrence.R
\name{occurrence_ui}
\alias{occurrence_ui}
\title{Co-Occurrence UI}
\usage{
occurrence_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\value{
A nav item that may be passed to a nav container
(e.g. \code{\link[bslib:navset]{bslib::navset_tab()}}).
}
\description{
Co-Occurrence UI
}
\seealso{
\code{\link[=occurrence_server]{occurrence_server()}}

Other count data modules: 
\code{\link{bertin_server}()},
\code{\link{bertin_ui}()},
\code{\link{count_server}()},
\code{\link{count_ui}()},
\code{\link{diversity_alpha_server}()},
\code{\link{diversity_alpha_ui}()},
\code{\link{diversity_beta_server}()},
\code{\link{diversity_beta_ui}()},
\code{\link{diversity_docs_ui}()},
\code{\link{diversity_server}()},
\code{\link{diversity_ui}()},
\code{\link{occurrence_server}()}
}
\concept{count data modules}
\keyword{internal}
