% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_server}
\alias{pca_server}
\title{Principal Components Analysis Server}
\usage{
pca_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{data.frame}.}
}
\value{
A reactive \code{\link[dimensio:PCA-class]{dimensio::PCA}} object.
}
\description{
Principal Components Analysis Server
}
\seealso{
\code{\link[=pca_ui]{pca_ui()}}

Other multivariate analysis modules: 
\code{\link{ca_server}()},
\code{\link{ca_ui}()},
\code{\link{multivariate_server}()},
\code{\link{multivariate_ui}()},
\code{\link{pca_ui}()}
}
\concept{multivariate analysis modules}
\keyword{internal}
