% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_ui}
\alias{pca_ui}
\title{Principal Components Analysis UI}
\usage{
pca_ui(id, center = TRUE, scale = TRUE, help = NULL)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}

\item{center}{A \code{\link{logical}} scalar: should the variables be shifted to be
zero centered?}

\item{scale}{A \code{\link{logical}} scalar: should the variables be scaled to unit
variance?}

\item{help}{A \code{\link{character}} string giving a short help text.}
}
\value{
A nav item that may be passed to a nav container
(e.g. \code{\link[bslib:navset]{bslib::navset_tab()}}).
}
\description{
Principal Components Analysis UI
}
\seealso{
\code{\link[=pca_server]{pca_server()}}

Other multivariate analysis modules: 
\code{\link{ca_server}()},
\code{\link{ca_ui}()},
\code{\link{multivariate_server}()},
\code{\link{multivariate_ui}()},
\code{\link{pca_server}()}
}
\concept{multivariate analysis modules}
\keyword{internal}
