% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_server}
\alias{prepare_server}
\title{Prepare Data Server}
\usage{
prepare_server(id, detect = function(...) TRUE, select = TRUE, demo = NULL)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{detect}{A predicate \code{\link{function}} used to select columns.}

\item{select}{A \code{\link{logical}} scalar: should all detected columns be selected?}

\item{demo}{A \code{\link{character}} string specifying the name of a dataset (see
\code{\link[=import_server]{import_server()}}).}
}
\value{
A reactive \code{\link{data.frame}}.
}
\description{
Prepare Data Server
}
\seealso{
\code{\link[=prepare_ui]{prepare_ui()}}

Other generic modules: 
\code{\link{import_server}()},
\code{\link{import_ui}()},
\code{\link{prepare_ui}()}
}
\concept{generic modules}
\keyword{internal}
