% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data.R
\name{prepare_ui}
\alias{prepare_ui}
\title{Prepare Data UI}
\usage{
prepare_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Prepare Data UI
}
\seealso{
\code{\link[=prepare_server]{prepare_server()}}

Other generic modules: 
\code{\link{data_server}()},
\code{\link{data_ui}()},
\code{\link{import_server}()},
\code{\link{import_ui}()},
\code{\link{missing_server}()},
\code{\link{missing_ui}()},
\code{\link{prepare_server}()}
}
\concept{generic modules}
\keyword{internal}
