% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{selectize_ui}
\alias{selectize_ui}
\title{Updatable Select List}
\usage{
selectize_ui(id, label = "Choose", multiple = FALSE)
}
\arguments{
\item{id}{A \code{\link{character}} string specifying the namespace.}
}
\value{
A select list control that can be added to a UI definition
(see \code{\link[shiny:selectInput]{shiny::selectizeInput()}}).
}
\description{
Updatable Select List
}
\seealso{
\code{\link[=update_selectize_colnames]{update_selectize_colnames()}}, \code{\link[=update_selectize_rownames]{update_selectize_rownames()}}
}
\keyword{internal}
