% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_seriate.R
\name{seriate_server}
\alias{seriate_server}
\title{Seriate Server}
\usage{
seriate_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{data.frame} (typically returned by \code{\link[=data_server]{data_server()}}).}
}
\value{
A reactive \code{\link[kairos:PermutationOrder-class]{kairos::AveragePermutationOrder}} object.
}
\description{
Seriate Server
}
\seealso{
\code{\link[=seriate_ui]{seriate_ui()}}

Other chronology modules: 
\code{\link{seriate_ui}()}
}
\concept{chronology modules}
\keyword{internal}
