% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_seriate.R
\name{seriate_ui}
\alias{seriate_ui}
\title{Seriate UI}
\usage{
seriate_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Seriate UI
}
\seealso{
\code{\link[=seriate_server]{seriate_server()}}

Other chronology modules: 
\code{\link{seriate_server}()}
}
\concept{chronology modules}
\keyword{internal}
