% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_summary.R
\name{summary_server}
\alias{summary_server}
\title{Data Summary Server}
\usage{
summary_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{data.frame} (typically returned by \code{\link[=import_server]{import_server()}}).}
}
\description{
Data Summary Server
}
\seealso{
\code{\link[=summary_ui]{summary_ui()}}

Other statistic modules: 
\code{\link{chi2_server}()},
\code{\link{chi2_ui}()},
\code{\link{summary_ui}()}
}
\concept{statistic modules}
\keyword{internal}
