% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_summary.R
\name{summary_ui}
\alias{summary_ui}
\title{Data Summary UI}
\usage{
summary_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Data Summary UI
}
\seealso{
\code{\link[=summary_server]{summary_server()}}

Other statistic modules: 
\code{\link{chi2_server}()},
\code{\link{chi2_ui}()},
\code{\link{summary_server}()}
}
\concept{statistic modules}
\keyword{internal}
