% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_ternary.R
\name{ternary_server}
\alias{ternary_server}
\title{Ternary Plot Server}
\usage{
ternary_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{data.frame}.}
}
\description{
Ternary Plot Server
}
\seealso{
\code{\link[=ternary_ui]{ternary_ui()}}

Other plot modules: 
\code{\link{ternary_ui}()}
}
\concept{plot modules}
\keyword{internal}
