% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary.R
\name{ternary_ui}
\alias{ternary_ui}
\title{Ternary Plot UI}
\usage{
ternary_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Ternary Plot UI
}
\seealso{
\code{\link[=ternary_server]{ternary_server()}}

Other plot modules: 
\code{\link{ternary_server}()}
}
\concept{plot modules}
\keyword{internal}
