% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis-internal.R
\name{theme_ui}
\alias{theme_ui}
\title{Bootstrap Theme}
\usage{
theme_ui(...)
}
\arguments{
\item{...}{Extra parameters to be passed to \code{\link[bslib:bs_theme]{bslib::bs_theme()}}.}
}
\description{
Bootstrap Theme
}
\seealso{
\code{\link[bslib:bs_theme]{bslib::bs_theme()}}
}
\keyword{internal}
