% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_interval.R
\name{time_interval_ui}
\alias{time_interval_ui}
\title{Time Intervals UI}
\usage{
time_interval_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\value{
A nav item that may be passed to a nav container
(e.g. \code{\link[bslib:navset]{bslib::navset_tab()}}).
}
\description{
Time Intervals UI
}
\seealso{
\code{\link[=time_interval_server]{time_interval_server()}}

Other chronology modules: 
\code{\link{aoristic_server}()},
\code{\link{aoristic_ui}()},
\code{\link{seriate_server}()},
\code{\link{seriate_ui}()},
\code{\link{time_interval_server}()}
}
\concept{chronology modules}
\keyword{internal}
