% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{update_checkbox_colnames}
\alias{update_checkbox_colnames}
\title{Update a Checkbox Group with Column Names}
\usage{
update_checkbox_colnames(id, x, exclude = reactive(NULL), select = TRUE)
}
\arguments{
\item{id}{A \code{\link{character}} string specifying the namespace.}

\item{x}{A reactive \code{matrix}-like object.}

\item{exclude}{A reactive \code{\link{character}} vector of values to be excluded from
choices.}

\item{select}{A \code{\link{logical}} scalar: should all choices be selected?}
}
\value{
A reactive \code{\link{character}} vector of column names.

Side effect: change the value of a checkbox group on the client.
}
\description{
Update a Checkbox Group with Column Names
}
\seealso{
\code{\link[=checkbox_ui]{checkbox_ui()}}
}
\keyword{internal}
