% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{update_selectize_rownames}
\alias{update_selectize_rownames}
\title{Update a Select List with Row Names}
\usage{
update_selectize_rownames(
  id,
  x,
  exclude = reactive(NULL),
  select = FALSE,
  placeholder = TRUE
)
}
\arguments{
\item{id}{A \code{\link{character}} string specifying the namespace.}

\item{x}{A reactive \code{matrix}-like object.}

\item{exclude}{A reactive \code{\link{character}} vector of values to be excluded from
choices.}

\item{select}{A \code{\link{logical}} scalar: should all choices be selected?}

\item{placeholder}{A \code{\link{logical}} scalar: should a placeholder be added?}
}
\value{
A reactive \code{\link{character}} vector of row names.

Side effect: change the value of a select list on the client.
}
\description{
Update a Select List with Row Names
}
\seealso{
\code{\link[=selectize_ui]{selectize_ui()}}
}
\keyword{internal}
