% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis-internal.R
\name{update_selectize_variables}
\alias{update_selectize_variables}
\title{Update a Select List with Column Names}
\usage{
update_selectize_variables(
  id,
  x,
  find = NULL,
  use = NULL,
  selected = NULL,
  preserve = TRUE,
  none = TRUE,
  server = TRUE
)
}
\arguments{
\item{id}{A \code{\link{character}} string specifying the namespace (must match
\code{\link[=selectize_ui]{selectize_ui()}}).}

\item{x}{A reactive \code{matrix}-like object.}

\item{find}{A predicate \code{\link{function}} for column selection
(see \code{\link[arkhe:detect]{arkhe::detect()}}).}

\item{selected}{A \code{\link{character}} vector specifying the initially selected
value(s).}

\item{preserve}{A \code{\link{logical}} scalar: should existing selection be preserved
on update?}

\item{none}{A \code{\link{logical}} scalar: should a placeholder be added as the first
element?}

\item{server}{A \code{\link{logical}} scalar: should server-side selectize be used?}
}
\value{
A reactive \code{\link{character}} vector of column names.

Side effect: change the value of a select input on the client.
}
\description{
Update a Select List with Column Names
}
\seealso{
\code{\link[=selectize_ui]{selectize_ui()}}
}
\keyword{internal}
