% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis-internal.R
\name{updateSelectValues}
\alias{updateSelectValues}
\title{Update a Select List with a Vector}
\usage{
updateSelectValues(
  id,
  x,
  exclude = reactive({
     NULL
 }),
  preserve = TRUE,
  none = TRUE,
  server = TRUE
)
}
\arguments{
\item{id}{A \code{\link{character}} string specifying the namespace (must match
\code{\link[=selectize_ui]{selectize_ui()}}).}

\item{x}{A reactive \code{\link{character}} vector.}

\item{exclude}{A reactive \code{\link{character}} vector of values to exclude.}

\item{preserve}{A \code{\link{logical}} scalar: should existing selection be preserved
on update?}

\item{none}{A \code{\link{logical}} scalar: should a placeholder be added as the first
element?}

\item{server}{A \code{\link{logical}} scalar: should server-side selectize be used?}
}
\value{
A reactive \code{\link{character}} vector of column names.
}
\description{
Update a Select List with a Vector
}
\seealso{
\code{\link[=selectize_ui]{selectize_ui()}}
}
\keyword{internal}
