% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.single.trial.GA.r
\name{view.single.trial.GA}
\alias{view.single.trial.GA}
\title{View single trial with Final Kinematics}
\usage{
view.single.trial.GA(tr, datatraj, datakin, details = F, frame = T)
}
\description{
View single trial with Final Kinematics
}
