% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.functions.r
\name{view.single.trial.grasp3d}
\alias{view.single.trial.grasp3d}
\title{interactive 3d plot}
\usage{
view.single.trial.grasp3d(tr, datatraj, datakin, entireMov = T)
}
\description{
interactive 3d plot
}
