% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.functions.r
\name{view.single.trial.projections}
\alias{view.single.trial.projections}
\title{display the trajectory of the reaching movement on the three major projection planes}
\usage{
view.single.trial.projections(tr, datatraj, datakin)
}
\description{
display the trajectory of the reaching movement on the three major projection planes
}
