% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.functions.r
\name{view.summary.trials}
\alias{view.summary.trials}
\title{view.summary.trials}
\usage{
view.summary.trials(trials = NULL, dataset, grid.size = 20, frame = T,
  print = T, res = 28)
}
\description{
view.summary.trials
}
