% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda_outliers.R
\name{coda_outliers_ui}
\alias{coda_outliers_ui}
\title{Compositional Data Outliers UI}
\usage{
coda_outliers_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Compositional Data Outliers UI
}
\seealso{
\code{\link[=coda_outliers_server]{coda_outliers_server()}}

Other coda modules: 
\code{\link{coda_outliers_server}()},
\code{\link{coda_plot_server}()},
\code{\link{coda_plot_ui}()},
\code{\link{coda_server}()},
\code{\link{coda_summary_server}()},
\code{\link{coda_summary_ui}()},
\code{\link{coda_ui}()},
\code{\link{logratio_server}()},
\code{\link{logratio_ui}()}
}
\concept{coda modules}
\keyword{internal}
