% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_server}
\alias{count_server}
\title{Count Data Server}
\usage{
count_server(id, demo = NULL)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{demo}{A \code{\link{character}} string specifying the name of a dataset (see
\code{\link[=import_server]{import_server()}}).}
}
\value{
A reactive \code{\link{data.frame}}.
}
\description{
Count Data Server
}
\seealso{
\code{\link[=count_ui]{count_ui()}}

Other count data modules: 
\code{\link{bertin_server}()},
\code{\link{bertin_ui}()},
\code{\link{count_ui}()},
\code{\link{diversity_alpha_server}()},
\code{\link{diversity_alpha_ui}()},
\code{\link{diversity_beta_server}()},
\code{\link{diversity_beta_ui}()},
\code{\link{diversity_docs_ui}()},
\code{\link{diversity_server}()},
\code{\link{diversity_ui}()},
\code{\link{occurrence_server}()},
\code{\link{occurrence_ui}()}
}
\concept{count data modules}
\keyword{internal}
