% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_diversity_docs.R
\name{diversity_docs_ui}
\alias{diversity_docs_ui}
\title{Diversity Definitions UI}
\usage{
diversity_docs_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\value{
A nav item that may be passed to a nav container
(e.g. \code{\link[bslib:navset]{bslib::navset_tab()}}).
}
\description{
Diversity Definitions UI
}
\seealso{
Other count data modules: 
\code{\link{bertin_server}()},
\code{\link{bertin_ui}()},
\code{\link{diversity_alpha_server}()},
\code{\link{diversity_alpha_ui}()},
\code{\link{diversity_beta_server}()},
\code{\link{diversity_beta_ui}()},
\code{\link{occurrence_server}()},
\code{\link{occurrence_ui}()}
}
\concept{count data modules}
\keyword{internal}
