% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home.R
\name{footer_ui}
\alias{footer_ui}
\title{Footer UI}
\usage{
footer_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\value{
A \code{\link{list}} that can be converted into an HTML \verb{<footer>} tag
(see \code{\link[htmltools:builder]{htmltools::tags()}}).
}
\description{
Footer UI
}
\seealso{
\code{\link[=footer_server]{footer_server()}}

Other page modules: 
\code{\link{footer_server}()},
\code{\link{home_server}()},
\code{\link{home_ui}()}
}
\concept{page modules}
\keyword{internal}
