% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_config}
\alias{get_config}
\title{Read Configuration Values}
\usage{
get_config(
  app,
  config = Sys.getenv("KINESIS_CONFIG_ACTIVE", "default"),
  use_parent = TRUE
)
}
\arguments{
\item{app}{A \code{\link{character}} string specifying the Shiny application
to run (see \code{\link[=run_app]{run_app()}}).}

\item{config}{A \code{\link{character}} string specifying the name of configuration to
read from.}

\item{use_parent}{A \code{\link{logical}} scalar: should parent directories be scanned
for configuration files if the specified config file isn't found?}
}
\value{
A \code{\link{list}} of configuration values.
}
\description{
Read Configuration Values
}
\author{
N. Frerebeau
}
\keyword{internal}
