% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{get_config}
\alias{get_config}
\title{Read Configuration Values}
\usage{
get_config(
  config = Sys.getenv("JANUS_CONFIG_ACTIVE", "default"),
  use_parent = TRUE
)
}
\arguments{
\item{config}{A \code{\link{character}} string specifying the name of configuration to
read from.}

\item{use_parent}{A \code{\link{logical}} scalar: should parent directories be scanned
for configuration files if the specified config file isn't found?}

\item{app}{A \code{\link{character}} string specifying the Shiny application
configuration to get . Any unambiguous substring can be given.}
}
\value{
A \code{\link{list}} of configuration values.
}
\description{
Read Configuration Values
}
\author{
N. Frerebeau
}
\keyword{internal}
