% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{get_config}
\alias{get_config}
\title{Read Configuration Values}
\usage{
get_config(app, file = NA, active = NA, use_parent = TRUE)
}
\arguments{
\item{app}{A \code{\link{character}} string specifying the Shiny application
to run (see \code{\link[=run_app]{run_app()}}).}

\item{file}{A \code{\link{character}} string specifying the configuration file to
read from. If \code{NA} (the default), use the value of the
\code{KINESIS_CONFIG_FILE} environment variable ("\code{config.yml}" if the variable
does not exist). If \code{NULL}, use the build-in configuration file.}

\item{active}{A \code{\link{character}} string specifying the name of configuration to
read from. If \code{NA} (the default), use the value of the
\code{KINESIS_CONFIG_ACTIVE} environment variable ("\code{default}" if the variable
does not exist).}

\item{use_parent}{A \code{\link{logical}} scalar: should parent directories be scanned
for configuration files if the specified config file isn't found?}
}
\value{
A \code{\link{list}} of configuration values.
}
\description{
Read Configuration Values
}
\author{
N. Frerebeau
}
\keyword{internal}
