% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home.R
\name{home_ui}
\alias{home_ui}
\title{About UI}
\usage{
home_ui(id, package)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}

\item{package}{A \code{\link{character}} vector of package names to be cited.}
}
\value{
A nav item that may be passed to a nav container
(e.g. \code{\link[bslib:navset]{bslib::navset_tab()}}).
}
\description{
About UI
}
\seealso{
\code{\link[=home_server]{home_server()}}

Other page modules: 
\code{\link{footer_server}()},
\code{\link{footer_ui}()},
\code{\link{home_server}()}
}
\concept{page modules}
\keyword{internal}
