% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_server}
\alias{import_server}
\title{Import Data Server}
\usage{
import_server(id, demo = NULL)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{demo}{A \code{\link{character}} string specifying the name of a dataset from
\pkg{folio} or \pkg{datasets}.}
}
\value{
A reactive \code{\link{data.frame}}.
}
\description{
Import Data Server
}
\seealso{
\code{\link[=import_ui]{import_ui()}}

Other generic modules: 
\code{\link{import_ui}()},
\code{\link{prepare_server}()},
\code{\link{prepare_ui}()}
}
\concept{generic modules}
\keyword{internal}
