% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin.sgFilter.r
\name{kin.sgFilter}
\alias{kin.sgFilter}
\title{apply Savitzky-Golay filter}
\usage{
kin.sgFilter(x, p = 3, n = p + 3 - p\%\%2, m = 0, ts = 1)
}
\description{
This functon is exactly identical to signal::sgolayfilt
}
