% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin.sgFilter.r
\name{kin.sgFilter}
\alias{kin.sgFilter}
\title{apply Savitzky-Golay filter}
\usage{
kin.sgFilter(x, p = 4, m = 0, ts = 1)
}
\description{
This function is a wrapper for signal::sgolayfilt with a modification to handle NAs
}
