% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis-package.R
\docType{package}
\name{kinesis-package}
\alias{kinesis-package}
\alias{_PACKAGE}
\alias{kinesis}
\title{kinesis: 'Shiny' Applications for the 'tesselle' Packages}
\description{
A collection of 'shiny' applications for the 'tesselle' packages.
}
\details{
\tabular{ll}{
\strong{Package:} \tab kinesis \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 0.0.0.9000 \cr
\strong{License:} \tab GPL-3 \cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tesselle/kinesis}
  \item \url{https://packages.tesselle.org/kinesis/}
  \item Report bugs at \url{https://github.com/tesselle/kinesis/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
