% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis-package.R
\docType{package}
\name{kinesis-package}
\alias{kinesis-package}
\alias{kinesis}
\title{kinesis: 'shiny' Applications for the 'tesselle' Packages}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of 'shiny' applications for the 'tesselle' packages \url{https://www.tesselle.org/}. This package provides applications for archaeological data analysis and visualization. These mainly, but not exclusively, include applications for chronological modelling (e.g. matrix seriation, aoristic analysis) and count data analysis (e.g. diversity measures, compositional data analysis).
}
\details{
\tabular{ll}{
\strong{Version} \tab 0.2.0 \cr
\strong{License} \tab GPL-3 \cr
\strong{Zenodo DOI} \tab \doi{10.5281/zenodo.14645671} \cr
}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\section{Package options}{

\pkg{kinesis} uses the following \code{\link[=options]{options()}} to configure behavior:
\itemize{
\item \code{kinesis.workers}: an \code{\link{integer}} specifying the number of \R sessions
to be used for asynchronous (parallel) computing. Defaults to 1.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://codeberg.org/tesselle/kinesis}
  \item \url{https://packages.tesselle.org/kinesis/}
  \item \url{https://tesselle.r-universe.dev/kinesis}
  \item Report bugs at \url{https://codeberg.org/tesselle/kinesis/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Frerebeau \email{nicolas.frerebeau@u-bordeaux-montaigne.fr} (\href{https://orcid.org/0000-0001-5759-4944}{ORCID})

Other contributors:
\itemize{
  \item Brice Lebrun (\href{https://orcid.org/0000-0001-7503-8685}{ORCID}) (Logo designer) [artist]
  \item Université Bordeaux Montaigne (03pbgwk21) [funder]
  \item CNRS (02feahw73) [funder]
}

}
\keyword{internal}
