% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_mcd.R
\name{mcd_ui}
\alias{mcd_ui}
\title{MCD UI}
\usage{
mcd_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\value{
A nav item that may be passed to a nav container
(e.g. \code{\link[bslib:navset]{bslib::navset_tab()}}).
}
\description{
MCD UI
}
\seealso{
\code{\link[=mcd_server]{mcd_server()}}

Other ceramic data modules: 
\code{\link{mcd_server}()}
}
\concept{ceramic data modules}
\keyword{internal}
