% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_footer.R
\name{module_footer_server}
\alias{module_footer_server}
\title{Footer Server}
\usage{
module_footer_server(id, cite = "janus")
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{cite}{A \code{\link{character}} string giving the name of the package to be
cited.}
}
\description{
Footer Server
}
\seealso{
\code{\link[=module_footer_ui]{module_footer_ui()}}

Other server modules: 
\code{\link{module_ca_server}()},
\code{\link{module_import_server}()},
\code{\link{module_seriate_server}()},
\code{\link{module_settings_server}()}
}
\concept{server modules}
\keyword{internal}
