% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_footer.R
\name{module_footer_server}
\alias{module_footer_server}
\title{Footer Server}
\usage{
module_footer_server(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\description{
Footer Server
}
\seealso{
\code{\link[=module_footer_ui]{module_footer_ui()}}

Other server modules: 
\code{\link{module_chi_server}()},
\code{\link{module_count_server}()},
\code{\link{module_header_server}()},
\code{\link{module_home_server}()},
\code{\link{module_import_server}()},
\code{\link{module_missing_server}()},
\code{\link{module_multivar_server}()},
\code{\link{module_prepare_server}()},
\code{\link{module_seriate_server}()},
\code{\link{module_summary_server}()}
}
\concept{server modules}
\keyword{internal}
