% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_import.R
\name{module_import_server}
\alias{module_import_server}
\title{Import Data Server}
\usage{
module_import_server(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}
}
\value{
A reactive \code{data.frame}.
}
\description{
Import Data Server
}
\seealso{
\code{\link[=module_import_ui]{module_import_ui()}}

Other server modules: 
\code{\link{module_chi_server}()},
\code{\link{module_count_server}()},
\code{\link{module_footer_server}()},
\code{\link{module_header_server}()},
\code{\link{module_home_server}()},
\code{\link{module_missing_server}()},
\code{\link{module_multivar_server}()},
\code{\link{module_pca_server}()},
\code{\link{module_prepare_server}()},
\code{\link{module_seriate_server}()},
\code{\link{module_summary_server}()},
\code{\link{module_ternary_server}()}
}
\concept{server modules}
\keyword{internal}
