% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_import.R
\name{module_import_ui}
\alias{module_import_ui}
\title{Import Data UI}
\usage{
module_import_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\description{
Import Data UI
}
\seealso{
\code{\link[=module_import_server]{module_import_server()}}

Other UI modules: 
\code{\link{module_footer_ui}()},
\code{\link{module_header_ui}()},
\code{\link{module_home_ui}()},
\code{\link{module_missing_ui}()},
\code{\link{module_multivar_ui}()},
\code{\link{module_prepare_ui}()},
\code{\link{module_seriate_ui}()},
\code{\link{module_summary_ui}()}
}
\concept{UI modules}
\keyword{internal}
