% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_settings_server.R
\name{module_settings_server}
\alias{module_settings_server}
\title{Settings Server}
\usage{
module_settings_server(input, output, session, user_settings)
}
\arguments{
\item{input}{provided by \pkg{Shiny}.}

\item{output}{provided by \pkg{Shiny}.}

\item{session}{provided by \pkg{Shiny}.}

\item{user_settings}{A \code{\link[shiny:reactiveValues]{shiny::reactiveValues()}} list.}
}
\description{
Settings Server
}
\seealso{
\code{\link[=module_settings_ui]{module_settings_ui()}}

Other server modules: 
\code{\link{module_ca_server}()},
\code{\link{module_footer_server}()},
\code{\link{module_import_server}()},
\code{\link{module_seriate_server}()},
\code{\link{module_visualize_server}()}
}
\concept{server modules}
