% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate.R
\name{multivariate_ui}
\alias{multivariate_ui}
\title{Mutlivariate Analysis UI}
\usage{
multivariate_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\value{
A tab that can be passed to \code{\link[shiny:tabsetPanel]{shiny::tabsetPanel()}}.

A navigation container (see \code{\link[bslib:navset]{bslib::navset_card_pill()}}).
}
\description{
Mutlivariate Analysis UI
}
\seealso{
\code{\link[=multivariate_server]{multivariate_server()}}

Other multivariate analysis modules: 
\code{\link{ca_server}()},
\code{\link{ca_ui}()},
\code{\link{multivariate_server}()},
\code{\link{pca_server}()},
\code{\link{pca_ui}()}
}
\concept{multivariate analysis modules}
\keyword{internal}
