% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{output_plot}
\alias{output_plot}
\title{Plot UI}
\usage{
output_plot(id, ..., tools = NULL, title = "Card title")
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}

\item{...}{Further parameters to be passed to \code{\link[shiny:plotOutput]{shiny::plotOutput()}}.}

\item{tools}{A (list of) input elements.}
}
\description{
Plot UI
}
\seealso{
Other widgets: 
\code{\link{render_plot}()}
}
\concept{widgets}
\keyword{internal}
