% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_server}
\alias{prepare_server}
\title{Prepare Data Server}
\usage{
prepare_server(id, choose = function(...) TRUE, select = function(...) TRUE)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{choose}{A predicate \code{\link{function}} used to select columns.}

\item{select}{A predicate \code{\link{function}} used to select columns.}
}
\value{
A reactive \code{data.frame}.
}
\description{
Prepare Data Server
}
\seealso{
\code{\link[=prepare_ui]{prepare_ui()}}

Other generic modules: 
\code{\link{import_server}()},
\code{\link{import_ui}()},
\code{\link{prepare_ui}()}
}
\concept{generic modules}
\keyword{internal}
