% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data.R
\name{prepare_server}
\alias{prepare_server}
\title{Prepare Data Server}
\usage{
prepare_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{data.frame} (typically returned by
\code{\link[=import_server]{import_server()}}).}
}
\value{
A reactive \code{data.frame}.
}
\description{
Prepare Data Server
}
\seealso{
\code{\link[=prepare_ui]{prepare_ui()}}

Other generic modules: 
\code{\link{data_server}()},
\code{\link{data_ui}()},
\code{\link{import_server}()},
\code{\link{import_ui}()},
\code{\link{missing_server}()},
\code{\link{missing_ui}()},
\code{\link{prepare_ui}()}
}
\concept{generic modules}
\keyword{internal}
