% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_ui}
\alias{prepare_ui}
\title{Prepare Data UI}
\usage{
prepare_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\value{
A nav item that may be passed to a nav container
(e.g. \code{\link[bslib:navset]{bslib::navset_tab()}}).
}
\description{
Prepare Data UI
}
\seealso{
\code{\link[=prepare_server]{prepare_server()}}

Other generic modules: 
\code{\link{import_server}()},
\code{\link{import_ui}()},
\code{\link{prepare_server}()}
}
\concept{generic modules}
\keyword{internal}
