% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plot.R
\name{render_plot}
\alias{render_plot}
\title{Plot Server}
\usage{
render_plot(id, x, width = "auto", height = NULL)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive recorded plot to be saved (see \code{\link[grDevices:recordplot]{grDevices::recordPlot()}}).}

\item{width, height}{Height and width specification (see \code{\link[shiny:renderPlot]{shiny::renderPlot()}}).}
}
\description{
Plot Server
}
\seealso{
Other plot modules: 
\code{\link{output_plot}()}
}
\concept{plot modules}
\keyword{internal}
