% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{render_plot}
\alias{render_plot}
\title{Plot Server}
\usage{
render_plot(id, x, ...)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{\link{function}} recording the plot.}

\item{...}{Further parameters to be passed to \code{\link[shiny:renderPlot]{shiny::renderPlot()}}.}
}
\value{
No return value, called for side effects.
}
\description{
Plot Server
}
\seealso{
Other widgets: 
\code{\link{output_plot}()}
}
\concept{widgets}
\keyword{internal}
