% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_app}
\alias{run_app}
\title{Run an App}
\usage{
run_app(app = c("seriation"), options = list(launch.browser = interactive()))
}
\arguments{
\item{app}{A \code{\link{character}} string specifying the Shiny application
to run (see details). Any unambiguous substring can be given.}

\item{options}{A \code{\link{list}} of named options that should be passed to the
\code{\link[shiny:runApp]{shiny::runApp()}} call.}
}
\value{
A \pkg{shiny} application object.
}
\description{
A wrapper for \code{\link[shiny:shinyApp]{shiny::shinyAppDir()}}.
}
\details{
\tabular{ll}{
\strong{Application name} \tab  \strong{Keyword} \cr
Matrix Seriation \tab \code{seriation} \cr
}
}
\examples{
\dontrun{
run_app("seriation")
}
}
\author{
N. Frerebeau
}
\concept{shiny apps}
