% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_app}
\alias{run_app}
\title{Run an App}
\usage{
run_app(app = c("seriate", "ca"), browser = TRUE, display = "auto")
}
\arguments{
\item{app}{A \code{\link{character}} string specifying the Shiny application
to run (see details). Any unambiguous substring can be given.}

\item{browser}{A \code{\link{logical}} scalar: should the app be run in
the browser?}

\item{display}{A \code{\link{character}} string specifying the mode in which
to display the application (see \code{\link[shiny:runApp]{shiny::runApp()}}).}
}
\value{
A \pkg{shiny} application object.
}
\description{
A wrapper for \code{\link[shiny:shinyApp]{shiny::shinyAppDir()}}.
}
\details{
\tabular{ll}{
\strong{Application name} \tab  \strong{Keyword} \cr
Matrix Seriation \tab \code{seriate} \cr
Correspondence Analysis \tab \code{ca} \cr
}
}
\examples{
\dontrun{
run_app("seriate")
}
}
\seealso{
Other shiny: 
\code{\link{deploy}()}
}
\author{
N. Frerebeau
}
\concept{shiny}
