% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter_server}
\alias{scatter_server}
\title{Scatter Plot Server}
\usage{
scatter_server(id, x)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's
UI function.}

\item{x}{A reactive \code{\link{data.frame}}.}
}
\value{
No return value, called for side effects.
}
\description{
Scatter Plot Server
}
\seealso{
\code{\link[=scatter_ui]{scatter_ui()}}

Other plot modules: 
\code{\link{graphics_server}()},
\code{\link{graphics_ui}()},
\code{\link{scatter_ui}()},
\code{\link{ternary_server}()},
\code{\link{ternary_ui}()}
}
\concept{plot modules}
\keyword{internal}
