% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter_ui}
\alias{scatter_ui}
\title{Scatter Plot UI}
\usage{
scatter_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\value{
A nav item that may be passed to a nav container
(e.g. \code{\link[bslib:navset]{bslib::navset_tab()}}).
}
\description{
Scatter Plot UI
}
\seealso{
\code{\link[=scatter_server]{scatter_server()}}

Other plot modules: 
\code{\link{graphics_server}()},
\code{\link{graphics_ui}()},
\code{\link{scatter_server}()},
\code{\link{ternary_server}()},
\code{\link{ternary_ui}()}
}
\concept{plot modules}
\keyword{internal}
