% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriate.R
\name{seriate_ui}
\alias{seriate_ui}
\title{CA Seriation UI}
\usage{
seriate_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\value{
A nav item that may be passed to a nav container
(e.g. \code{\link[bslib:navset]{bslib::navset_tab()}}).
}
\description{
CA Seriation UI
}
\seealso{
\code{\link[=seriate_server]{seriate_server()}}

Other chronology modules: 
\code{\link{aoristic_server}()},
\code{\link{aoristic_ui}()},
\code{\link{seriate_server}()},
\code{\link{time_interval_server}()},
\code{\link{time_interval_ui}()}
}
\concept{chronology modules}
\keyword{internal}
