% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary.R
\name{ternary_ui}
\alias{ternary_ui}
\title{Ternary Plot UI}
\usage{
ternary_ui(id)
}
\arguments{
\item{id}{A \code{\link{character}} vector to be used for the namespace.}
}
\value{
A nav item that may be passed to a nav container
(e.g. \code{\link[bslib:navset]{bslib::navset_tab()}}).
}
\description{
Ternary Plot UI
}
\seealso{
\code{\link[=ternary_server]{ternary_server()}}

Other plot modules: 
\code{\link{graphics_server}()},
\code{\link{graphics_ui}()},
\code{\link{scatter_server}()},
\code{\link{scatter_ui}()},
\code{\link{ternary_server}()}
}
\concept{plot modules}
\keyword{internal}
