% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinesis-internal.R
\name{theme_ui}
\alias{theme_ui}
\title{Bootstrap Theme}
\usage{
theme_ui(version = "5", ...)
}
\arguments{
\item{version}{A \code{\link{character}} string specifying the major version of
Bootstrap to use.}

\item{...}{Extra parameters to be passed to \code{\link[bslib:bs_theme]{bslib::bs_theme()}}.}
}
\value{
Returns a \code{\link[sass:sass_layer]{sass::sass_bundle()}} object (see \code{\link[bslib:bs_theme]{bslib::bs_theme()}}).
}
\description{
Bootstrap Theme
}
\keyword{internal}
